refactor: replace createBasicVd() to module
